fix(jzero): fix gen swagger

+++ internal/gen/genswagger/gen.go
@@ -25,34 +25,38 @@ var (
 func Gen(_ *cobra.Command, _ []string) error {
 	wd, _ := os.Getwd()
 
-	mainApiFile, isDelete, err := gen.GetMainApiFilePath(ApiDir)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if isDelete {
-			_ = os.Remove(mainApiFile)
-		}
-	}()
-
-	if !pathx.FileExists(Dir) {
+	if pathx.FileExists(ApiDir) {
 		_ = os.MkdirAll(Dir, 0o755)
-	}
-
-	// gen swagger by desc/api
-	if mainApiFile != "" {
-		apiFile := fmt.Sprintf("%s.swagger.json", gen.GetApiServiceName(ApiDir))
-		cmd := exec.Command("goctl", "api", "plugin", "-plugin", "goctl-swagger=swagger -filename "+apiFile+" --schemes http", "-api", mainApiFile, "-dir", Dir)
-		resp, err := cmd.CombinedOutput()
+		mainApiFile, isDelete, err := gen.GetMainApiFilePath(ApiDir)
 		if err != nil {
-			return errors.Wrap(err, strings.TrimRight(string(resp), "\r\n"))
+			return err
 		}
-		if strings.TrimRight(string(resp), "\r\n") != "" {
-			fmt.Println(strings.TrimRight(string(resp), "\r\n"))
+		defer func() {
+			if isDelete {
+				_ = os.Remove(mainApiFile)
+			}
+		}()
+
+		if !pathx.FileExists(Dir) {
+			_ = os.MkdirAll(Dir, 0o755)
+		}
+
+		// gen swagger by desc/api
+		if mainApiFile != "" {
+			apiFile := fmt.Sprintf("%s.swagger.json", gen.GetApiServiceName(ApiDir))
+			cmd := exec.Command("goctl", "api", "plugin", "-plugin", "goctl-swagger=swagger -filename "+apiFile+" --schemes http", "-api", mainApiFile, "-dir", Dir)
+			resp, err := cmd.CombinedOutput()
+			if err != nil {
+				return errors.Wrap(err, strings.TrimRight(string(resp), "\r\n"))
+			}
+			if strings.TrimRight(string(resp), "\r\n") != "" {
+				fmt.Println(strings.TrimRight(string(resp), "\r\n"))
+			}
 		}
 	}
 
 	if pathx.FileExists(ProtoDir) {
+		_ = os.MkdirAll(Dir, 0o755)
 		protoDirFile, err := os.ReadDir(ProtoDir)
 		if err != nil {
 			return err
